+++ History.md
@@ -1,5 +1,7 @@
 # master
 
+* Fix sprockets integration (#17)
+
 # 0.8.2
 
 * Fixed "Unexpected token return" ([issue #10](https://github.com/jfirebaugh/skim/issues/10))